all_requirements.txt
adium-theme-ubuntu==0.3.4
apptools==4.3.0
attrs==15.2.0
autobahn==0.10.3
autopep8==1.2.4
awscli==1.11.2
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.1
botocore==1.4.60
certifi==2017.11.5
chardet==2.3.0
colorama==0.3.7
colorlog==2.7.0
configobj==5.0.6
configparser==3.5.0
cryptography==1.2.3
cycler==0.9.0
Cython==0.24.1
dask==0.11.0
decorator==4.0.6
dnspython==1.12.0
docutils==0.12
easydict==1.6
entrypoints==0.2.3
enum34==1.1.2
envisage==4.4.0
filterpy==1.0.0
flake8==3.2.1
functools32==3.2.3.post2
futures==3.0.5
Glances==2.10
gyp==0.1
h5py==2.6.0
html5lib==1.0b10
idna==2.0
importmagic==0.1.7
ipaddress==1.0.16
ipdb==0.10.1
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.1.0
ipywidgets==7.0.4
jedi==0.9.0
Jinja2==2.10
jmespath==0.9.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
leveldb==0.194
lmdb==0.93
lz4==0.7.0
MarkupSafe==1.0
matplotlib==1.5.1
mayavi==4.4.3
mccabe==0.5.3
meld==3.14.2
mistune==0.8.1
mpi4py==1.3.1
msgpack-python==0.4.6
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
nose==1.3.7
notebook==5.2.1
num2words==0.5.5
numpy==1.11.0
PAM==0.4.2
pandas==0.18.1
pandocfilters==1.4.2
pathlib2==2.1.0
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.1.2
pkg-resources==0.0.0
progressbar2==3.32.1
prompt-toolkit==1.0.7
protobuf==3.1.0.post1
psutil==5.2.2
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycodestyle==2.2.0
pycrypto==2.6.1
pydot==1.2.3
pyface==4.5.2
pyflakes==1.3.0
Pygments==2.1
pygobject==3.20.0
pygraphviz==1.3.1
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyserial==3.0.1
Pyste==0.9.10
python-dateutil==2.4.2
python-gflags==3.0.7
python-snappy==0.5
python-utils==2.2.0
pytz==2014.10
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
rsa==3.4.2
s3transfer==0.1.7
scikit-image==0.12.3
scikit-learn==0.18
scipy==0.17.0
service-identity==16.0.0
simplegeneric==0.8.1
SimPy==2.3.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
terminado==0.6
testpath==0.3.1
toolz==0.8.0
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.1
traits==4.5.0
traitsui==4.5.1
trollius==2.0.1
Twisted==16.0.0
txaio==1.0.0
unity-lens-photos==1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.7
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xdot==0.6
xgboost==0.6
yapf==0.15.0
zope.interface==4.1.3
